+++ devices/wda/types.go
@@ -3,15 +3,12 @@ package wda
 import (
 	"net/http"
 	"strings"
-	"sync"
 	"time"
 )
 
 type WdaClient struct {
 	baseURL    string
 	httpClient *http.Client
-	sessionId  string
-	mu         sync.Mutex
 }
 
 func NewWdaClient(hostPort string) *WdaClient {
@@ -25,6 +22,9 @@ func NewWdaClient(hostPort string) *WdaClient {
 		baseURL: baseURL,
 		httpClient: &http.Client{
 			Timeout: 60 * time.Second,
+			Transport: &http.Transport{
+				DisableKeepAlives: true,
+			},
 		},
 	}
 }
@@ -37,17 +37,3 @@ type TapAction struct {
 	Button   int    `json:"button"`
 }
 
-type PointerParameters struct {
-	PointerType string `json:"pointerType"`
-}
-
-type Pointer struct {
-	Type       string            `json:"type"`
-	ID         string            `json:"id"`
-	Parameters PointerParameters `json:"parameters"`
-	Actions    []TapAction       `json:"actions"`
-}
-
-type ActionsRequest struct {
-	Actions []Pointer `json:"actions"`
-}